+++ CHANGELOG.md
@@ -15,6 +15,8 @@
     * Bumped tldextract version to 3.1.0
     * Bumped chardet version to 4.0.0
     * Bumped requests to 2.25.1
+* Added input validation for `phishbench.initialize` 
+
     
 ## Bug Fixes